Renames package from nomic.game.manager.* to nomic.*
