+++ src/main/kotlin/nomic/domain/entities/Entity.kt
